// src/data/enums.ts
export enum ActionKind {
  SetIsInit = "setIsInit",
  SetCards = "setCards",
  SetExistingCards = "setExistingCards",

  SetCorrectNum = "setCorrectNum",
  SetCurrentNumber = "setCurrentNumber",
  SetErrorSelect = "setErrorSelect",
  SetSuccessSelect = "setSuccessSelect",
  SetIsAnswered = "setIsAnswered",
  SetIsFavorited = "setIsFavorited",
  SetIsSumbitted = "setIsSumbitted"
}